refactor: move auth folder to routes
