+++ src/app.module.ts
@@ -12,7 +12,7 @@ import { FavsModule } from './routes/favs/favs.module';
 import { CustomLoggerService } from './сommon/logger/logger.service';
 import { CustomExceptionFilter } from './сommon/exception-filter/exeption.filter';
 import { LoggerMiddleware } from './сommon/logger/logger.middleware';
-import { AuthModule } from './auth/auth.module';
+import { AuthModule } from './routes/auth/auth.module';
 
 @Module({
   imports: [
