+++ src/routes/auth/auth.service.ts
@@ -1,8 +1,8 @@
 import { ForbiddenException, Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import * as bcrypt from 'bcrypt';
-import { CreateUserDto } from '../routes/user/dto/create-user.dto';
-import { UserService } from '../routes/user/user.service';
+import { CreateUserDto } from '../user/dto/create-user.dto';
+import { UserService } from '../user/user.service';
 import { IPayload } from './models/payload.interface';
 
 @Injectable()
