+++ src/routes/auth/jwt.strategy.ts
