+++ src/routes/favs/favs.controller.ts
@@ -9,7 +9,7 @@ import {
   Post,
   UseGuards,
 } from '@nestjs/common';
-import { JwtGuard } from 'src/auth/jwt.guard';
+import { JwtGuard } from '../auth/jwt.guard';
 import { FavsService } from './favs.service';
 
 @Controller('favs')
