+++ src/routes/track/track.module.ts
@@ -6,7 +6,7 @@ import { AlbumModule } from 'src/routes/album/album.module';
 import { FavsModule } from 'src/routes/favs/favs.module';
 import TrackEntity from './entities/track.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from 'src/auth/auth.module';
+import { AuthModule } from '../auth/auth.module';
 
 @Module({
   imports: [
