+++ src/routes/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from 'src/auth/auth.module';
+import { AuthModule } from '../auth/auth.module';
 import UserEntity from './entities/user.entity';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
